Add spec coverage to hammertime class
Cursor did not do well with adding coverage
to the hammertime class. Many examples had
to be pending, in part due to dependency on
spec invocation ordering.

Having complete specs is probably not necessary
for the limited functionality of the gem, but
having them would allow some refactoring, and
if done correctly provide a template for
similar work in the future.